2.5: Separate module step10
completed
